Replace Block Factory with Blockly Factory.  Save Block Factory in blockfactory_old for safety. Keep icon.png and blocks.js back into blockfactory rather than pointing to blockfactory_old.
