+++ demos/blockfactory/app_controller.js
