+++ demos/blockfactory/block_library_controller.js
