+++ demos/blockfactory/block_library_view.js
