+++ demos/blockfactory/block_option.js
