+++ demos/blockfactory/factory_utils.js
