+++ demos/blockfactory/standard_categories.js
