+++ demos/blockfactory/workspacefactory/wfactory_controller.js
