+++ demos/blockfactory/workspacefactory/wfactory_generator.js
