+++ demos/blockfactory/workspacefactory/wfactory_init.js
