+++ demos/blockfactory/workspacefactory/wfactory_view.js
